\name{hdl}
\docType{data}
\alias{hdl}
\title{Mus Musculus HDL QTL data from Leduc et. al. (2012)
} 
\description{HDL QTL data was obtained from a F2 inner-cross between inbred MRL/MpJ and SM/J strains of mice.
}
\usage{
data(hdl)
}
\format{The data set \code{hdl} is a data frame of 280 observations of 15 variables: genotype data (genotype states at 5 SNP markers) and phenotype data (HDL levels and normalized expression values of 10 genes).Genotypes are of class \code{factor} and phenotypes are of class \code{numeric}.
}

\details{
HDL QTL data was obtained from a F2 inner-cross between inbred MRL/MpJ and SM/J strains of mice.
}
\source{
Leduc MS, Blair RH, Verdugo RA, Tsaih SW, Walsh K, Churchill GA, Paigen B.(2012). "Using bioinformatics and systems genetics to dissect HDL-cholesterol genetics in an MRL/MpJ x SM/J intercross." \emph{J Lipid Res.}, 6, 1163-75.
}

\keyword{datasets}
