\name{hdl}
\docType{data}
\alias{hdl}
\alias{hdlgeno}
\alias{hdlpheno}
\title{Mus Musculus HDL QTL data from Leduc et. al. (2012)
} 
\description{HDL QTL data was obtained from a F2 inner-cross between inbred MRL/MpJ and SM/J strains of mice.
}
\usage{
data(hdl)
}
\format{The data set \code{hdl} contains the following variables:
\itemize{
\item{hdlgeno}{  a data frame of 280 observations of 5 variables. All column vectors are of class factor}
\item{hdlpheno}{ a data frame of 280 observations of 10 variables. All column vectors are of class numeric }
}
}
\details{
HDL QTL data was obtained from a F2 inner-cross between inbred MRL/MpJ and SM/J strains of mice.
}
\source{
Leduc MS, Blair RH, Verdugo RA, Tsaih SW, Walsh K, Churchill GA, Paigen B.(2012). "Using bioinformatics and systems genetics to dissect HDL-cholesterol genetics in an MRL/MpJ x SM/J intercross." \emph{J Lipid Res.}, 6, 1163-75.
}

\keyword{datasets}
